requirements.txt
opencv-python
tensorflow-gpu==2.2.0
keras==2.3.1
numpy
Pillow
matplotlib